+++ package.json
@@ -13,6 +13,7 @@
     "clsx": "^2.0.0",
     "next": "12.3.1",
     "next-auth": "^4.23.1",
+    "qs": "^6.11.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "swr": "^2.2.2"
