+++ pages/api/auth/[...nextauth].js
@@ -9,6 +9,19 @@ export const authOptions = {
       clientSecret: process.env.NEXT_PUBLIC_GOOGLE_CLIENT_SECRET,
     }),
   ],
+  pages: {
+    signIn: "/",
+    error: "/",
+  },
+
+  callbacks: {
+    async signIn({ account, profile }) {
+      if (account.provider === "google") {
+        return profile.email_verified && profile.email.endsWith("@jala.tech");
+      }
+      return true; // Do different verification for other providers that don't have `email_verified`
+    },
+  },
 };
 
 export default NextAuth(authOptions);
